cleanup from refactor
