+++ .github/workflows/release.yml
@@ -1,5 +1,5 @@
 ---
-name: Release
+name: release
 
 on:
   push:
@@ -14,18 +14,40 @@ jobs:
       matrix:
         include:
           - os: ubuntu-latest
-            artifact_name: beerus_linux_x86_64.tar.gz
-            asset_name: beerus_linux_x86_64.tar.gz
+            artifact_name: beerus-linux-x86_64.tar.gz
+            asset_name: beerus-linux-x86_64.tar.gz
           - os: macos-latest
-            artifact_name: beerus_darwin_x86_64.tar.gz
-            asset_name: beerus_darwin_x86_64.tar.gz
+            artifact_name: beerus-darwin-x86_64.tar.gz
+            asset_name: beerus-darwin-x86_64.tar.gz
 
     steps:
-    - uses: actions/checkout@v3
+      - name: Setup Toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          
+      - name: Cache Deps
+        uses: actions/cache@v3
+        with:
+          path: |
+            ~/.cargo/bin/
+            ~/.cargo/registry/index/
+            ~/.cargo/registry/cache/
+            ~/.cargo/git/db/
+            target/
+          key: ${{ matrix.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
+      
+    - name: Checkout Repo
+      uses: actions/checkout@v3
+
     - name: Build
       run: cargo build --release --locked
+
     - name: Create Tarball
-      run: tar -C target/release -czvf ${{ matrix.artifact_name }} beerus beerus-rpc
+      run: >
+        tar -C target/release -czvf ${{ matrix.artifact_name }}
+        beerus beerus-rpc
+
     - name: Upload binaries to release
       uses: svenstaro/upload-release-action@v2
       with:
