+++ README.md
@@ -21,34 +21,27 @@
 
 </div>
 
-![](docs/images/beerus.gif)
-
-<details open="open">
+<details>
 <summary>Table of Contents</summary>
 
 - [Report a Bug](#report-a-bug)
 - [Request a Feature](#request-a-feature)
-- [Report a bug](#report-a-bug-1)
-- [Request a feature](#request-a-feature-1)
 - [About](#about)
   - [Built With](#built-with)
 - [Architecture](#architecture)
-- [Simple usage overview](#simple-usage-overview)
+  - [Simple usage overview](#simple-usage-overview)
 - [Getting Started](#getting-started)
   - [Prerequisites](#prerequisites)
   - [Installation](#installation)
     - [Beerusup](#beerusup)
     - [Environment variables](#environment-variables)
-  - [Configuration](#configuration)
-- [Usage](#usage)
-  - [CLI](#cli)
-    - [Ethereum](#ethereum)
-      - [Query balance](#query-balance)
-    - [StarkNet](#starknet)
-      - [Query contract view](#query-contract-view)
-      - [Query get storage at](#query-get-storage-at)
-  - [API](#api)
+    - [Configuration](#configuration)
+    - [Build](#build)
+    - [Test](#test)
+    - [Debug](#debug)
 - [Roadmap](#roadmap)
+- [Report a bug](#report-a-bug-1)
+- [Request a feature](#request-a-feature-1)
 - [Support](#support)
 - [Project assistance](#project-assistance)
 - [Contributing](#contributing)
@@ -62,31 +55,11 @@
 
 ---
 
-## Report a bug
-
-If you find a bug, please open a
-[GitHub issue](https://github.com/starknet-exploration/beerus/issues/new?assignees=&labels=bug&template=01_BUG_REPORT.md&title=bug%3A+)!
-
-## Request a feature
-
-To request a new feature, please open an issue following
-[this template](https://github.com/starknet-exploration/beerus/issues/new?assignees=&labels=enhancement&template=02_FEATURE_REQUEST.md&title=feat%3A+).
-
 ## About
 
-> Beerus is a StarkNet Light Client inspired by and using
-> [helios](https://github.com/a16z/helios/). The goal is to provide a simple and
-> easy to use client to query StarkNet state and interact with contracts.
-
-<details>
-<summary>Screenshots</summary>
-<br>
-
-|                             Screenshot 1                              |                              Screenshot 2                              |
-| :-------------------------------------------------------------------: | :--------------------------------------------------------------------: |
-| <img src="docs/images/screenshot.png" title="Home Page" width="100%"> | <img src="docs/images/screenshot.png" title="Login Page" width="100%"> |
-
-</details>
+Beerus is a StarkNet Light Client inspired by and using
+[helios](https://github.com/a16z/helios/). The goal is to provide a simple and
+easy to use client to query StarkNet state and interact with contracts.
 
 ### Built With
 
@@ -100,7 +73,7 @@ Here is a high level overview of the architecture of Beerus.
 
 [![Beerus architecture](docs/images/beerus-architecture-v1.0.png)](docs/images/beerus-architecture-v1.0.png)
 
-## Simple usage overview
+### Simple usage overview
 
 Here is a simple overview of how Beerus work. The example is for querying a
 storage value of a StarkNet contract.
@@ -133,66 +106,7 @@ For StarkNet node for the moment you can use Infura but soon
 will be implemented in Pathfinder nodes, and so will these nodes be working as
 well.
 
-Copy the `.env.example` file to a `.env` file and populate each variable (they
-are all mandatory except for `ETHEREUM_NETWORK` which defaults to `"goerli"`:
-
-```bash
-cp examples/.env.example .env
-```
-
-Meanwhile you can just use unit tests to dev.
-
-```bash
-cargo test --all
-```
-
-Run binary:
-
-```bash
-source .env && cargo run --bin beerus-rpc
-```
-
-Run binary verbose:
-
-```bash
-source .env && RUST_LOG=info cargo run --bin beerus-rpc
-```
-
-Build from source:
-
-```bash
-cargo build --all --release
-```
-
-The binaries will be located in `target/release/`. Specifically, the binary for
-the CLI is `target/release/beerus` and the binary for the API is
-`target/release/beerus-rpc`.
-
-Specify the environment variables and run the binary.
-
-For example to query the state root of the StarkNet network using the CLI:
-
-```bash
-source .env && ./target/release/beerus starknet query-state-root
-# Should output something like:
-# 3018024614248707887030376849268409283849258987090518998455999582305744756580
-```
-
-To do the same using the API:
-
-```bash
-source .env && ./target/release/beerus
-# Then go to http://127.0.0.1:8000/starknet/state/root
-# Should output something like:
-```
-
-```json
-{
-  "state_root": "3018024614248707887030376849268409283849258987090518998455999582305744756580"
-}
-```
-
-### Configuration
+#### Configuration
 
 Beerus is configurable through environment variables.
 
@@ -215,53 +129,37 @@ Here is the list of all the available environment variables:
 | ETHEREUM_NETWORK           | goerli        | The Ethereum network to use. Can be one of `mainnet`, `goerli`.                                             |
 | ETHEREUM_EXECUTION_RPC_URL | No            | Ethereum execution layer RPC URL (must be an Ethereum provider that supports the eth_getProof endpoint)     |
 | ETHEREUM_CONSENSUS_RPC_URL | No            | Ethereum consensus layer RPC URL (must be a consensus node that supports the light client beacon chain api) |
-| STARKNET_RPC_URL           | No            | StarkNet RPC URL                                                                                            |
+| STARKNET_CORE_CONTRACT_ADDRESS | No            | location of core contracts |
+| STARKNET_RPC_URL           | No            | StarkNet RPC URL |
 
-## Usage
-
-### CLI
-
-```bash
-Usage: beerus [OPTIONS] <COMMAND>
-
-Commands:
-  ethereum  Ethereum related subcommands
-  help      Print this message or the help of the given subcommand(s)
-
-Options:
-  -c, --config <FILE>  Sets a custom config file
-  -h, --help           Print help information
-  -V, --version        Print version information
-```
-
-#### Ethereum
-
-##### Query balance
+Copy the `.env.example` file to a `.env` file and populate each variable (they
+are all mandatory except for `ETHEREUM_NETWORK` which defaults to `"goerli"`:
 
 ```bash
-beerus ethereum query-balance --address 0x00000000219ab540356cBB839Cbe05303d7705Fa
-# 2011.286832686010020640 ETH
+cp examples/.env.example .env
 ```
-
-#### StarkNet
-
-##### Query contract view
 
 ```bash
-beerus starknet query-contract --address 0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7 --selector 0x1e888a1026b19c8c0b57c72d63ed1737106aa10034105b980ba117bd0c29fe1 --calldata 0x00,0x01
-[FieldElement { inner: 0x0000000000000000000000000000000000000000000000000000000000000000 }, FieldElement { inner: 0x0000000000000000000000000000000000000000000000000000000000000000 }]
+source .env
 ```
 
-##### Query get storage at
+#### Build
 
 ```bash
-beerus starknet query-get-storage-at --address 0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7 --key 0x341c1bdfd89f69748aa00b5742b03adbffd79b8e80cab5c50d91cd8c2a79be1
-298305742194
+cargo build --release
 ```
 
-### API
+#### Test
 
-> **[TODO]**
+```bash
+cargo test
+```
+
+#### Debug
+
+```bash
+RUST_LOG=info cargo run --bin beerus-rpc
+```
 
 ## Roadmap
 
@@ -273,6 +171,16 @@ a list of proposed features (and known issues).
 - [Top Bugs](https://github.com/starknet-exploration/beerus/issues?q=is%3Aissue+is%3Aopen+label%3Abug+sort%3Areactions-%2B1-desc)
   (Add your votes using the 👍 reaction)
 - [Newest Bugs](https://github.com/starknet-exploration/beerus/issues?q=is%3Aopen+is%3Aissue+label%3Abug)
+
+## Report a bug
+
+If you find a bug, please open a
+[GitHub issue](https://github.com/starknet-exploration/beerus/issues/new?assignees=&labels=bug&template=01_BUG_REPORT.md&title=bug%3A+)!
+
+## Request a feature
+
+To request a new feature, please open an issue following
+[this template](https://github.com/starknet-exploration/beerus/issues/new?assignees=&labels=enhancement&template=02_FEATURE_REQUEST.md&title=feat%3A+).
 
 ## Support
